:recycle: Refactor: 카테고리 관련 컴포넌트 폴더 분리
- components/category 폴더 생성
- feat: useModal.confirm 콜백 함수 작성
