+++ src/application/hooks/common/useModal.tsx
@@ -8,9 +8,12 @@ const useModal = () => {
   const { show: showToast, close } = useToast();
 
   const confirm = useCallback(
-    (message: string) => {
+    (message: string, { ...rest }: { onSuccess: () => void }) => {
       close();
-      showToast({ content: <Confirm title={message} />, type: 'modal' });
+      showToast({
+        content: <Confirm title={message} {...rest} onReject={close} />,
+        type: 'modal',
+      });
     },
     [close, showToast],
   );
