+++ src/components/scrap/Toast/index.ts
@@ -1,5 +1,5 @@
+export { default as SelectCategoryToast } from '../../category/SelectCategory';
 export { default as CreateScrapToast } from './CreateScrap';
 export { default as DeleteScrapToast } from './DeleteScrap';
-export { default as SelectCategoryToast } from './SelectCategory';
 export { default as TypedCompleteToast } from './TypedComplete';
 export { default as TypedDetailToast } from './TypedDetail';
